Fix the broadcast errors, and restore flow typing
I agree that the code was cleaner, but the extra layer of indirection through the template meant that flow couldn't type-check most of the code. I've put back just enough of the old code to get type-checking to work.

With the type-checking working again, it was obvious what the broadcast error was.