Change refillServers to deplete serverList before calling getServers
Since we only need 3 servers but we ask for 8, have refillServers go through all 8 before it calls getServers again. This ensures that all 8 servers have a chance to be scored before pinging the cache again. Othewise, we could end up with some unused servers in the list that never get tried.

Implemented by simply holding the serverList in the engineState object instead of a local variable in refillServers.